Addition of BusinessException in QuoteDetailService and other changes in the QuoteDetail API REST. Other changes IN CustomerService

+++ src/main/java/com/dulsystems/mta/service/QuoteDetailService.java
@@ -3,6 +3,7 @@ package com.dulsystems.mta.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import com.dulsystems.mta.bean.QuoteBean;
@@ -12,6 +13,7 @@ import com.dulsystems.mta.bean.ResponseBean;
 import com.dulsystems.mta.bean.VehicleBean;
 import com.dulsystems.mta.dao.QuoteDetailDao;
 import com.dulsystems.mta.dao.VehicleDao;
+import com.dulsystems.mta.exception.BusinessException;
 
 @Service
 public class QuoteDetailService implements IQuoteDetailService{
@@ -30,11 +32,10 @@ public class QuoteDetailService implements IQuoteDetailService{
 		qb = quoteDetailDao.searchQuoteById(id);
 		if(qb != null) {
 			response.setCode("OK");
-			response.setMessage("Si existe la cotizacion");
+			response.setMessage("Consulta realizada");
 			response.setQb(qb);
 		}else {
-			response.setCode("BAD00");
-			response.setMessage("No existe la cotizacion");
+			throw new BusinessException("E-SERVICE-DAO",HttpStatus.BAD_REQUEST,"No existe la cotizacion");
 		}
 		return response;
 	}
@@ -46,15 +47,13 @@ public class QuoteDetailService implements IQuoteDetailService{
 		if(vb != null) {
 			if(quoteDetailDao.executeSaveQuote(request, vb) == true) {
 				response.setCode("OK");
-				response.setMessage("Se registro correctamente");
+				response.setMessage("Se guardo el registro");
 			
 			}else{
-				response.setCode("BAD00");
-				response.setMessage("No se pudo registrar correctamente");
+				throw new BusinessException("E-SERVICE-DAO",HttpStatus.BAD_REQUEST,"No se pudo guardar el registro");
 			}
 		}else {
-			response.setCode("BAD01");
-			response.setMessage("El vehiculo no existe");
+			throw new BusinessException("E-SERVICE-DAO_VALIDATIONS",HttpStatus.BAD_REQUEST,"No existe el vehiculo");
 		}
 		return response;
 	}
@@ -62,19 +61,21 @@ public class QuoteDetailService implements IQuoteDetailService{
 	@Override
 	public ResponseBean executeUpdateQuoteById(RequestBean request) {
 		ResponseBean response = new ResponseBean();
-		VehicleBean vb = vehicleDao.searchVehicleByPlate(request.getVehiclePlate());
-		if(vb != null) {
-			if(quoteDetailDao.executeUpdateQuoteById(request, vb) == true) {
-				response.setCode("OK");
-				response.setMessage("Se actualizo correctamente");
-			
-			}else{
-				response.setCode("BAD00");
-				response.setMessage("No se pudo actualizar correctamente");
+		if(quoteDetailDao.searchQuoteById(request.getQuoteId())!=null) {
+			VehicleBean vb = vehicleDao.searchVehicleByPlate(request.getVehiclePlate());
+			if(vb != null) {
+				if(quoteDetailDao.executeUpdateQuoteById(request, vb) == true) {
+					response.setCode("OK");
+					response.setMessage("Se actualizo el registro");
+				
+				}else{
+					throw new BusinessException("E-SERVICE-DAO",HttpStatus.BAD_REQUEST,"No se pudo actualizar el registro");
+				}
+			}else {
+				throw new BusinessException("E-SERVICE-DAO_VALIDATIONS",HttpStatus.BAD_REQUEST,"No existe el vehiculo");
 			}
 		}else {
-			response.setCode("BAD01");
-			response.setMessage("El vehiculo no existe");
+			throw new BusinessException("E-SERVICE-DAO_VALIDATIONS",HttpStatus.BAD_REQUEST,"No existe la cotizacion que quieres actualizar");
 		}
 		return response;
 	}
@@ -82,36 +83,37 @@ public class QuoteDetailService implements IQuoteDetailService{
 	@Override
 	public ResponseBean removeQuoteById(Integer id) {
 		ResponseBean response = new ResponseBean();
-		if(quoteDetailDao.removeQuoteById(id)) {
-			response.setCode("OK");
-			response.setMessage("Se elimino correctamente");
-			
-		}else{
-			response.setCode("BAD00");
-			response.setMessage("No se pudo eliminar correctamente");
+		if(quoteDetailDao.searchQuoteById(id)!=null) {
+			if(quoteDetailDao.removeQuoteById(id)) {
+				response.setCode("OK");
+				response.setMessage("Se elimino el registro");
+				
+			}else{
+				response.setCode("BAD00");
+				response.setMessage("No se pudo eliminar correctamente");
+				throw new BusinessException("E-SERVICE-DAO",HttpStatus.BAD_REQUEST,"No se pudo eliminar el registro");
+			}
+		}else {
+			throw new BusinessException("E-SERVICE-DAO_VALIDATIONS",HttpStatus.BAD_REQUEST,"No se elimino porque la cotizacion no existe");
 		}
         return response;
 	}
 	
 	//SERVICES FOR DETAILS OF A QUOTE
 	@Override
-	public ResponseBean searchQuoteDetailsById(Integer quoteDetail) {
+	public ResponseBean searchQuoteDetailsByQuoteId(Integer quoteDetail) {
 		ResponseBean response = new ResponseBean();
-		QuoteBean qb = quoteDetailDao.searchQuoteById(quoteDetail);
-		
-		if(qb != null) {
-			List<QuoteDetailBean> lqdb = quoteDetailDao.searchQuoteDetailById(quoteDetail);
-			if(lqdb != null) {
+		if(quoteDetailDao.searchQuoteById(quoteDetail) != null) {
+			List<QuoteDetailBean> lqdb = quoteDetailDao.searchQuoteDetailsByQuoteId(quoteDetail);
+			if(lqdb.size() > 0) {
 				response.setCode("OK");
-				response.setMessage("Si hay detalles de la cotizacion");
+				response.setMessage("Consulta realizada");
 				response.setLQdb(lqdb);
 			}else {
-				response.setCode("BAD00");
-				response.setMessage("No hay detalles de la cotizacion");
+				throw new BusinessException("E-SERVICE-DAO",HttpStatus.BAD_REQUEST,"No existen detalles de la cotizacion");
 			}
 		}else {
-			response.setCode("BAD01");
-			response.setMessage("No existe la cotizacion");
+			throw new BusinessException("E-SERVICE-DAO_VALIDATIONS",HttpStatus.BAD_REQUEST,"No existe la cotizacion");
 		}
 		return response;
 	}
@@ -119,53 +121,48 @@ public class QuoteDetailService implements IQuoteDetailService{
 	@Override
 	public ResponseBean executeSaveQuoteDetails(RequestBean request) {
 		ResponseBean response = new ResponseBean();
-		//QuoteBean qb = quoteDao.searchQuoteById(request.getQuoteId());
-		if(true/*qb != null*/) {
-			if(quoteDetailDao.executeSaveQuoteDetail(request, /*qb,*/ request.getLqdb()).length > 0) {
+		if(quoteDetailDao.searchQuoteById(request.getLqdb().get(0).getQuoteDetailIdFk()) != null) {
+			if(quoteDetailDao.executeSaveQuoteDetails(request, request.getLqdb()).length > 0) {
 				response.setCode("OK");
-				response.setMessage("Se registraron los detalles de la cotizacion correctamente");
-			
+				response.setMessage("Se guardaron los detalles de la cotizacion");
 			}else{
-				response.setCode("BAD00");
-				response.setMessage("No se pudieron actualizar los detalles correctamente");
+				throw new BusinessException("E-SERVICE-DAO",HttpStatus.BAD_REQUEST,"No se pudieron guardar los detalles de la cotizacion");
 			}
 		}else {
-			response.setCode("BAD01");
-			response.setMessage("El vehiculo no existe");
+			throw new BusinessException("E-SERVICE-DAO_VALIDATIONS",HttpStatus.BAD_REQUEST,"No existe la cotizacion");
 		}
 		return response;
 	}
 
 	@Override
-	public ResponseBean executeUpdateQuoteDetailsById(RequestBean request) {
+	public ResponseBean executeUpdateQuoteDetailsByQuoteId(RequestBean request) {
 		ResponseBean response = new ResponseBean();
-		//QuoteBean qb = quoteDao.searchQuoteById(request.getQuoteId());
-		if(true/*qb != null*/) {
-			if(quoteDetailDao.executeUpdateQuoteDetailById(request, /*qb,*/ request.getLqdb()).length > 0) {
+		if(quoteDetailDao.searchQuoteById(request.getLqdb().get(0).getQuoteDetailIdFk()) != null) {
+			if(quoteDetailDao.executeUpdateQuoteDetailsByQuoteId(request, request.getLqdb()).length > 0) {
 				response.setCode("OK");
-				response.setMessage("Se actualizaron los detalles de la cotizacion correctamente");
-			
+				response.setMessage("Se actualizaron los detalles de la cotizacion");
 			}else{
-				response.setCode("BAD00");
-				response.setMessage("No se pudieron actualizar los detalles correctamente");
+				throw new BusinessException("E-SERVICE-DAO",HttpStatus.BAD_REQUEST,"No se pudieron actualizar los detalles de la cotizacion");
 			}
 		}else {
-			response.setCode("BAD01");
-			response.setMessage("El vehiculo no existe");
+			throw new BusinessException("E-SERVICE-DAO_VALIDATIONS",HttpStatus.BAD_REQUEST,"No existe la cotizacion");
 		}
 		return response;
 	}
 
 	@Override
-	public ResponseBean removeQuoteDetailById(Integer quoteDetail) {
+	public ResponseBean removeQuoteDetailsByQuoteId(Integer quoteDetail) {
 		ResponseBean response = new ResponseBean();
-		if(quoteDetailDao.removeQuoteDetailById(quoteDetail)) {
-			response.setCode("OK");
-			response.setMessage("Se elimino correctamente");
-			
-		}else{
-			response.setCode("BAD00");
-			response.setMessage("No se pudo eliminar correctamente");
+		if(quoteDetailDao.searchQuoteById(quoteDetail) != null) {
+			if(quoteDetailDao.removeQuoteDetailsByQuoteId(quoteDetail)) {
+				response.setCode("OK");
+				response.setMessage("Se eliminaron los detalles de la cotizacion");
+				
+			}else{
+				throw new BusinessException("E-SERVICE-DAO",HttpStatus.BAD_REQUEST,"No se pudieron eliminar los detalles de la cotizacion");
+			}
+		}else {
+			throw new BusinessException("E-SERVICE-DAO_VALIDATIONS",HttpStatus.BAD_REQUEST,"No se elimino porque los detalles de la cotizacion no existen");
 		}
         return response;
 	}
